Additional cleanup

+++ sidplayfp/c64/VIC_II/lightpen.h
@@ -71,12 +71,12 @@ public:
     /**
      * Return the low byte of x coordinate.
      */
-    uint8_t getX() const { return lpx; }
+    uint8_t getX() const { return (uint8_t)lpx; }
 
     /**
      * Return the low byte of y coordinate.
      */
-    uint8_t getY() const { return lpy; }
+    uint8_t getY() const { return (uint8_t)lpy; }
 
     /**
      * Retrigger lightpen on vertical blank.
